+++ ICD.Connect.Routing.CrestronPro/DigitalMedia/DmNvx/Dm100xStrBase/AbstractDm100XStrBaseAdapterSettings.cs
@@ -7,13 +7,13 @@ namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmNvx.Dm100xStrBase
 {
 	public abstract class AbstractDm100XStrBaseAdapterSettings : AbstractDeviceSettings, IDm100XStrBaseAdapterSettings
 	{
-		private const string ETHERNET_ID_ELEMENT = "EthernetId";
+		private const string IPID_ELEMENT = "IPID";
 		private const string ENDPOINT_ID_ELEMENT = "EndpointId";
 		private const string DIRECTOR_ID_ELEMENT = "DirectorId";
 		private const string DOMAIN_ID_ELEMENT = "DomainId";
 
 		[CrestronByteSettingsProperty]
-		public byte? EthernetId { get; set; }
+		public byte? Ipid { get; set; }
 
 		public uint? EndpointId { get; set; }
 
@@ -34,7 +34,7 @@ namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmNvx.Dm100xStrBase
 		{
 			base.WriteElements(writer);
 
-			writer.WriteElementString(ETHERNET_ID_ELEMENT, StringUtils.ToString(EthernetId));
+			writer.WriteElementString(IPID_ELEMENT, StringUtils.ToString(Ipid));
 			writer.WriteElementString(ENDPOINT_ID_ELEMENT, StringUtils.ToString(EndpointId));
 			writer.WriteElementString(DIRECTOR_ID_ELEMENT, StringUtils.ToString(DirectorId));
 			writer.WriteElementString(DOMAIN_ID_ELEMENT, StringUtils.ToString(DomainId));
@@ -48,7 +48,7 @@ namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmNvx.Dm100xStrBase
 		{
 			base.ParseXml(xml);
 
-			EthernetId = XmlUtils.TryReadChildElementContentAsByte(xml, ETHERNET_ID_ELEMENT);
+			Ipid = XmlUtils.TryReadChildElementContentAsByte(xml, IPID_ELEMENT);
 			EndpointId = XmlUtils.TryReadChildElementContentAsUInt(xml, ENDPOINT_ID_ELEMENT);
 			DirectorId = XmlUtils.TryReadChildElementContentAsInt(xml, DIRECTOR_ID_ELEMENT);
 			DomainId = XmlUtils.TryReadChildElementContentAsUInt(xml, DOMAIN_ID_ELEMENT);